rpc/client: make HTTP client more configurable
Closes #6162
