+++ light/provider/http/http_test.go
@@ -54,7 +54,7 @@ func TestProvider(t *testing.T) {
 	chainID := genDoc.ChainID
 	t.Log("chainID:", chainID)
 
-	c, err := rpchttp.New(rpcAddr, "/websocket")
+	c, err := rpchttp.New(rpcAddr)
 	require.Nil(t, err)
 
 	p := lighthttp.NewWithClient(chainID, c)
